+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Automatic reload settings on change in the settings file. Not all settings are reloaded automatically, 
 but the most ones are. The ones that are not automatically reloaded: `useGpt`, `coqLspPath`, `parseFileOnInit`.
 - Added command that solves admits in selected region. Also added that command to the context menu (right click in the editor).
-
+- Fix toggle extension. 
 
 ### 1.4.1
 - Add a possibility to configure a single tactic solver.